Fix clang tidy warnings incrementally

+++ parse.c
@@ -474,7 +474,7 @@ static void jump_patch(Parser* parser, intmax_t offset) {
 
     const uint16_t u16_jump = (uint16_t)jump;
     LOG("patch jump=%hu\n", u16_jump);
-    const uint8_t b1 = (u16_jump >> 8);
+    const uint8_t b1 = (u16_jump >> (uint8_t)8);
     const uint8_t b2 = (uint8_t)u16_jump;
 
     parser->compiler->fn->chunk.opcodes[offset] = b1;
